per_line returns an iterator
The common case on these problems is to process each line once, one at a
time. For this purpose, we don't need to collect the lines into a `Vec`
and can be more efficient by keeping it as an iterator. Update the
`per_line` helper accordingly.

This does unfortunately make lifetimes more confusing, but that's what
we signed up for when we chose Rust.