Write out Rc::clone and Arc::clone
As suggested by the Rust Book, write `Rc::clone(&foo)` instead of
`foo.clone()` when `foo` is an `Rc`, to emphasize the cheap nature of
this clone.

I'm not sure what to do with calling clone on my own `TrackedState`
struct. It's cheap too, and I'm not sure the best way to communicate
this. Maybe I'll give it a method named `clone_cheaply` instead of
implementing `Clone` or something.